backend/requirements.txt
dj-database-url
django-ninja
django
python-dotenv
django-cors-headers
requests
psycopg2